fix(agent-gui): remove nested startup import waterfall
