+++ apps/desktop/src/renderer/src/features/workspace-workbench/ui/StandaloneAgentWindow.test.ts
@@ -117,7 +117,7 @@ test("standalone Agent handles task and app Agent launch requests", () => {
   );
   assert.match(
     standaloneWindowSource,
-    /<LazyDesktopAgentGUIWorkbenchBody[\s\S]*?prefillPromptBootstrapRequest=\{prefillPromptBootstrapRequest\}/
+    /<DesktopAgentGUIWorkbenchBody[\s\S]*?prefillPromptBootstrapRequest=\{prefillPromptBootstrapRequest\}/
   );
   assert.match(
     workbenchBodySource,
@@ -188,7 +188,22 @@ test("standalone Agent hides panel toggles until its content mounts", () => {
   );
   assert.match(
     standaloneWindowSource,
-    /<StandaloneAgentWindowContentReady onReady=\{handleContentReady\}>[\s\S]*?<LazyDesktopAgentGUIWorkbenchBody/
+    /<StandaloneAgentWindowContentReady onReady=\{handleContentReady\}>[\s\S]*?<DesktopAgentGUIWorkbenchBody/
+  );
+});
+
+test("standalone Agent loads its body with the route instead of adding a second lazy boundary", () => {
+  assert.match(
+    standaloneWindowSource,
+    /import \{ DesktopAgentGUIWorkbenchBody \} from "@renderer\/features\/workspace-agent\/ui\/DesktopAgentGUIWorkbenchBody\.tsx"/
+  );
+  assert.doesNotMatch(
+    standaloneWindowSource,
+    /LazyDesktopAgentGUIWorkbenchBody/
+  );
+  assert.doesNotMatch(
+    standaloneWindowSource,
+    /import\("@renderer\/features\/workspace-agent\/ui\/DesktopAgentGUIWorkbenchBody\.tsx"\)/
   );
 });
